refactor: 存储结构二维数组改为一维数组

+++ Timer.cpp
@@ -33,7 +33,7 @@ void Timer::StartTimer() {
     gettimeofday(&start, nullptr);
 }
 
-double Timer::EndTimer() {
+int64_t Timer::EndTimer() {
     gettimeofday(&end, nullptr);
     return GetElapsedTime();
 }
